% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_params.R
\name{default_model_params}
\alias{default_model_params}
\alias{model_params}
\title{Create model parameter specifications}
\usage{
default_model_params()

model_params(.inherit_from = NULL, VPmax = NULL, alpha = NULL,
  kN = NULL, mP = NULL, mp2 = NULL, Imax = NULL, kz = NULL,
  phiP = NULL, phiD = NULL, betaz = NULL, kNz = NULL, mz = NULL,
  mz2 = NULL, VD = NULL, mD = NULL, wmix = NULL, CtoChl = NULL,
  Pinit = NULL, Ninit = NULL, Zinit = NULL, Dsinit = NULL, kw = NULL,
  kc = NULL, tstep = NULL, nyears = NULL, flag_stn = NULL,
  flag_LI = NULL, flag_atten = NULL, flag_irrad = NULL,
  flag_PIcurve = NULL, flag_grazing = NULL, flag_outtype = NULL,
  flag_outfreq = NULL, flag_integ = NULL)
}
\arguments{
\item{VPmax}{VPmax(0) (2.5): max. rate photosynthesis, g C (g chl)-1 h-1}

\item{alpha}{alpha (0.15): initial slope of P-I curve, g C (g chl)-1 h-1 (W
m-2)-1}

\item{kN}{kN (0.85): half saturation constant for N uptake, mmol m-3}

\item{mP}{mP (0.02): phytoplankton mortality rate, linear d-1}

\item{mp2}{mp2 (0.025): phytoplankton mortality rate, quadratic, (mmol N
m-3)-1 d-1}

\item{Imax}{Imax (1.0): zooplankton max. ingestion, d-1}

\item{kz}{kz (0.86): zooplankton half sat. for intake, mmol N m-3}

\item{phiP}{phiP (0.67): zooplankton preference: P}

\item{phiD}{phiD (0.33): zooplankton preference: D}

\item{betaz}{betaz (0.69): zooplankton absorption efficiency: N}

\item{kNz}{kNz (0.75): zooplankton net production efficiency: N}

\item{mz}{mz (0.02): zooplankton mortality rate, linear, d-1}

\item{mz2}{mz2 (0.34): zooplankton moratlity, quadratic, (mmol N m-3)-1 d-1}

\item{VD}{VD (6.43): detritus sinking rate, m d-1}

\item{mD}{mD (0.06): deteritus remineralisation rate, d-1}

\item{wmix}{wmix (0.13): cross-thermocline mixing rate, m d-1}

\item{CtoChl}{CtoChl (75.0) carbon to chlorophyll ratio, g g-1}

\item{Pinit}{initial P, mmol N m-3}

\item{Ninit}{initial N, mmol m-3}

\item{Zinit}{initial Z, mmol N m-3}

\item{Dsinit}{inital Ds, mmol N m-3}

\item{kw}{light attenuation by water, m-1}

\item{kc}{light attenuation by phyto., (mmol N m-3)-1}

\item{tstep}{time step, day}

\item{nyears}{run duration, years (1 year= 365 days)}

\item{flag_stn}{station; 1=India (60N 20W), 2=Biotrans (47N 20W), 3=Kerfix
(50 40S 68 25E), 4=Papa (50N, 145W)}

\item{flag_LI}{1=numeric; 2=Evans & Parslow (1985); 3=Anderson (1993)}

\item{flag_atten}{1: kw+kcP (one layer for ML); 2: Anderson (1993) (spectral,
layers 0-5,5-23,>23 m}

\item{flag_irrad}{1=triangular over day; 2=sinusoidal over day}

\item{flag_PIcurve}{1=Smith fn; 2=exponential fn}

\item{flag_grazing}{(1) phihat_i = phi_i, (2) phihat_i = phi_i*P_i}

\item{flag_outtype}{output to files: 0=none, 1=last year only, 2=whole
simulation}

\item{flag_outfreq}{frequency for writing output: 1=every day, 0=every time
step}

\item{flag_integ}{choice of integration method: 0=Euler, 1=Runge Kutta 4}
}
\value{
A model_params object, which is a list of parameter values.
}
\description{
Creates a default parameter specification as distributed in the original file
"NPZD_parms.txt" and "NPZD_extras.txt". Values can be overridden from default
using default_model_params(), or overridden from an arbitrary parameter
specification using model_params().
}
\examples{
default_model_params()

}
