% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMPOWER.R
\name{EMPOWER}
\alias{EMPOWER}
\title{Run the EMPOWER model}
\usage{
EMPOWER(params = default_model_params(), directory = NULL)
}
\arguments{
\item{params}{A parameter set generated by \link{model_params}}

\item{directory}{An optional working directory in which to run the model}
}
\value{
An object of class EMPOWER_result
}
\description{
Run the EMPOWER model
}
