% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aphy.R
\name{FNaphy}
\alias{FNaphy}
\title{Anderson (1993) function for calculating daily photosynthesis}
\usage{
FNaphy(ss, ztop, zbottom, aphylast)
}
\arguments{
\item{ss}{Not sure}

\item{ztop}{not sure}

\item{zbottom}{not sure}

\item{aphylast}{not sure}
}
\value{
"acalc", whatever that is
}
\description{
Uses in-built piecewise light extinction and assumes sinusoidal irradiance during day
Water column divided into three layers: 0-5m, 5-23m and >23m
}
\keyword{internal}
