% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daylcalc.R
\name{FNdaylcalc}
\alias{FNdaylcalc}
\title{Calculation of day length as function of day of year and latitude}
\usage{
FNdaylcalc(jday, latradians)
}
\arguments{
\item{jday}{The julian day of the year}

\item{latradians}{Latitude, in radians}
}
\value{
Day length (in hours?)
}
\description{
Calculation of day length as function of day of year and latitude
}
\keyword{internal}
