% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_flux.R
\name{FNget_flux}
\alias{FNget_flux}
\title{Handle specification of the ecosystem model}
\usage{
FNget_flux(X)
}
\arguments{
\item{X}{The array that carries state variables flux[i,j], which is passed back to the
core code for integration}
}
\value{
A matrix flux[i,j], which is passed back to the core code for integration.
}
\description{
The terms in the differential equations are calculated and transferred to
matrix, flux[i,j], which is passed back to the core code for integration.
The user can define auxiliary variables, Y[i], that are written to output
files (state variables are automatically stored for output).
}
\references{
Anderson (1993)
Evans and Parslow (1985)
Eppley function of sea surface temperature
}
\keyword{internal}
