% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIcalcA93.R
\name{FNLIcalcA93}
\alias{FNLIcalcA93}
\title{Polynomial approximation of daily? phyotosynthesis}
\usage{
FNLIcalcA93(zdepth, Iin, Iout, kPARlay, alpha, Vp, daylnow, ahashnow)
}
\arguments{
\item{zdepth}{no idea}

\item{Iin}{no idea}

\item{Iout}{no idea}

\item{kPARlay}{no idea}

\item{alpha}{no idea}

\item{Vp}{no idea}

\item{daylnow}{no idea}

\item{ahashnow}{no idea}
}
\value{
Lim_I, probably a scalar
}
\description{
Photosynthesis calculated using polynomial approximation (Anderson, 1993)
This function is only used when Anderson (1993) is selected for calculating daily depth-integrated photosynthesis
}
\references{
Anderson (1993)
Platt et al. (1990)
}
\keyword{internal}
