% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIcalcEP85.R
\name{FNLIcalcEP85}
\alias{FNLIcalcEP85}
\title{Evans & Parslow (1985) function for calculating daily photosynthesis}
\usage{
FNLIcalcEP85(zdepth, Iin, Iout, kPARlay, alpha, Vp, daylnow)
}
\arguments{
\item{zdepth}{not sure}

\item{Iin}{not sure}

\item{Iout}{not sure}

\item{kPARlay}{not sure}

\item{alpha}{not sure}

\item{Vp}{not sure}

\item{daylnow}{not sure}
}
\value{
Lim_I (probably a scalar)
}
\description{
Light extinction using Beer's Law and assumed triangular irradiance during day
}
\references{
Evans & Parslow (1985)
}
\keyword{internal}
