% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIcalcNum.R
\name{FNLIcalcNum}
\alias{FNLIcalcNum}
\title{Calculation of daily photosynthesis}
\usage{
FNLIcalcNum(zdepth, Iin, Iout, kPARlay, alpha, Vp, daylnow, choiceirrad,
  choicePI)
}
\arguments{
\item{zdepth}{no idea}

\item{Iin}{no idea}

\item{Iout}{no idea}

\item{kPARlay}{no idea}

\item{alpha}{no idea}

\item{Vp}{no idea}

\item{daylnow}{no idea}

\item{choiceirrad}{no idea}

\item{choicePI}{no idea}
}
\value{
Lim_I (probably a scalar?)
}
\description{
numeric intergration (time) and analytic integration (depth)
Diel cycle of irradiance: choice of (1) triangular vs (2) sinusoidal (flag_irrad)
P-I curve: choice of (1) Smith fn vs (2) exponential fn  (flag_PIcurve)
numeric integration (time) uses nstepst+1 steps and trapezoidal rule for first and last steps
(the user may manually change nstepst in the code if so desired)
}
\keyword{internal}
