% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_core.R
\name{FNmodel_core}
\alias{FNmodel_core}
\title{Run the EMPOWER model core code}
\usage{
FNmodel_core()
}
\description{
This function is a function-wrapped version of EMPOWER.r, originally
distributed with Anderson et al. (2016). It returns nothing, instead writing
three files to the working directory: "out_aux.txt", "out_fluxes.txt", and
"out_statevars.txt". These files have no headers, and are actually
comma-separaed-values files (i.e., read using \link[utils]{read.csv}). The
function has been modified from the original to (1) not modify the global
environment, (2) not print anything unless specified, and (3) allow some
flexibility of inputs that would otherwise have to be specified by modifying
the code.
}
\references{
Anderson et al. (2016)
}
