% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noonparcalc.R
\name{FNnoonparcalc}
\alias{FNnoonparcalc}
\title{Calculation of noon irradiance}
\usage{
FNnoonparcalc(jday, latradians, clouds, e0)
}
\arguments{
\item{jday}{Julian day of the year}

\item{latradians}{Latitude (in radians)}

\item{clouds}{Clouds parameter}

\item{e0}{no idea}
}
\value{
Noon PAR, W m-2
}
\description{
Calculation of noon irradiance
}
\keyword{internal}
