% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_params.R
\name{print.model_params}
\alias{print.model_params}
\title{Print a model_params object}
\usage{
\method{print}{model_params}(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{ignored}
}
\value{
The object, invisibly
}
\description{
Print a model_params object
}
